refactor(theme): getDefaultConfig 迁移至 runtime 工具供两模式共用
把主题默认 app.config 生成函数从构建期 src/utils/defaults 移到 src/runtime/utils/theme-defaults，消除 runtime 反向依赖构建期 utils 的隐患（dist 运行时无法解析）；组件改用具名 useTheme 导入避免 barrel 拉取。
